Switch to nix-instantiate/nix-build approach
replaces materialisation - we do not need that any longer.
The important observation is that you can create multiple *.drv
files in one call to nix-instantiate if the nix expression
produces a list/set of derivations. Then, it is possible to
build these *.drvs independently, which we use for CI-like
behaviour.
There is a downside: You _need_ to do IFD-evaluation once
whenever any sources change (no materialisation) but the
upside is that we do not need to bother setting up
materialisation - which ends up being a pain, especially
for CI behaviour and related automation.